pyproject.toml
[project]
name = "fantasy-agent-eval"
version = "0.1.0"
description = "Tool-using NHL fantasy agent scaffold (Phase A)."
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["requests>=2.31.0", "python-dotenv>=1.0.1"]

[project.optional-dependencies]
dev = ["pytest>=7.4.0"]
openai = ["openai>=1.40.0"]
gemini = ["google-generativeai>=0.7.0"]
anthropic = ["anthropic>=0.32.0"]
llm = ["openai>=1.40.0", "google-generativeai>=0.7.0", "anthropic>=0.32.0"]

[tool.pytest.ini_options]
markers = ["network: hits live NHL API"]
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

